Should be a float, locally it is, but travis complains.

+++ spec/server_spec.rb
@@ -127,7 +127,7 @@ describe "Arango::Server" do
     end
 
     it 'time' do
-      expect(@server.time).to be_a(Float)
+      expect(@server.time).to be_a(Numeric)
     end
 
     it 'detailed_version' do
